Silence compiler warning from int/size_t mismatch
Amends 4757b93b0ef149a3564582a45589f731755c4236.

+++ tests/auto/network/access/qnetworkreply/tst_qnetworkreply.cpp
@@ -8053,7 +8053,7 @@ void tst_QNetworkReply::varyingCacheExpiry()
         u"http://localhost"_qs,
         u"http://localhost"_qs,
     };
-    for (int i = 0; i < std::size(urls); ++i)
+    for (size_t i = 0; i < std::size(urls); ++i)
         urls[i].setPort(servers[i].serverPort());
 
     // After the initial request is completed the connection is kept alive
